Fix(rerank): Correct relevanceScore setter to use double to avoid format exception - Changed setRelevanceScore parameter from int to double - Prevented IllegalFormatConversionException when formatting with %.4f

+++ spring-ai-rag/src/main/java/org/springframework/ai/rag/postretrieval/rerank/RerankResponse.java
@@ -30,7 +30,7 @@ public class RerankResponse {
         private int index;
 
         @JsonProperty("relevance_score")
-        private int relevanceScore;
+        private double relevanceScore;
 
         public int getIndex() {
             return index;
@@ -40,11 +40,11 @@ public class RerankResponse {
             this.index = index;
         }
 
-        public int getRelevanceScore() {
+        public double getRelevanceScore() {
             return relevanceScore;
         }
 
-        public void setRelevanceScore(int relevanceScore) {
+        public void setRelevanceScore(double relevanceScore) {
             this.relevanceScore = relevanceScore;
         }
     }
